pyproject.toml
[project]
name = "prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agent-framework>=1.0.0b251204",
    "azure-ai-inference>=1.0.0b9",
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.7.0b2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "semantic-kernel>=1.39.0",
    "streamlit>=1.52.1",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
